requirements.txt
# Core
pyttsx3>=2.90
faster-whisper>=1.0.0
sentence-transformers>=2.2.2
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0

# Optional - for enhanced LLM feedback (requires Ollama running locally)
# requests>=2.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
